api: remove UpdateContext, redraw automatically instead
This prevents any chance of a ImGui frame lingering for a whole timer
interval. That would be bad because it would prevent the backends from
re-rendering the previous frame's data in that time (eg. in response to
the window being resized).

Another way to achive the same goal would be to prevent opening a second
frame during the same timer cycle. That would enforce UpdateContext
being invoked exactly once per cycle.