fix a crash when opening a window more than 1 minute after the previous one in the same context [p=2639537]
Dear ImGui is not calling Viewport::onChanged() on the first frame of each window.
This causes FontList::setScale to not be called to initialize the texture.
The first frame of every window use the primary monitor's font texture because of this.
This commit prevents the texture manager from forgetting that texture (since the font list does not).