add a setting to disable hardware acceleration [p=2763726]
Windows users with Intel GPUs are experiencing random rendering artifacts (ocornut/imgui#3033).

Option is only available for Direct3D on Windows and OpenGL on macOS.

On Linux the only way to force a software renderer is by setting the
LIBGL_ALWAYS_SOFTWARE environment variable, which Mesa reads from
glxInitialize/eglInitialize which are per-display not per-window,
so dynamically overriding that per context is not possible.