fix a potential crash when removing textures
`j` must be resynchronized with `i` after cookie crumbs are deleted to match the
removed textures (when cmd.type was Remove). `i` is decremented when starting a
new Remove command (when wantCmd is set to Remove).

This led to subsequent comparisons being off-by-one and commands on textures
that are out-of-bounds.

Gating the resynchronization to when `cmd.type == TextureCmd::Remove` would be
an equivalent fix as it cannot happen on the same iteration that decrements `i`.